Refactor the sgen command to allow for sub commands. Fixed a bug where the output path wasn't being used to render the sdk client
