+++ pkg/config/config.go
@@ -4,5 +4,5 @@ type Config struct {
 	File         string
 	Output       string
 	DebugEnabled bool
-	Version      bool
+	Version      string
 }
